requirements.txt
yfinance>=0.2.40
pandas>=2.0
numpy>=1.24
statsmodels>=0.14
matplotlib>=3.7
pyarrow>=14.0
python-dateutil>=2.8
